PS-22: Correct tests and separate CureModifier
